Fixed grayscale issue on IE
I had issue with images not being greyed-out in IE. The grayscale property does not work on IE. I had to add two separate photos and use the hover option to make the second photo appear. I added all photo in the img folder. Did that for all logos in clients logos and awards section logos