Fixed the caption in the news cards
I did a few fixes of the header for IE and the font too.
Managed to get the caption in the newscards to stay fixed to the card by changing the positioning of the ancestor element to relative and then attach the caption with absolute positioning to it. It works on all browsers and viewpoints