Newscards fixes
I added some code to fix issues in the newscard section. There was a gap on the bottom of the card just underneath the date that had to be removed. I changed some of my HTML and made it look the same as the main website. Fixed for all screen sizes and browsers