Responsiveness of culture element
I added width and margins to the culture tab in order to make it responsive to different screen sizes. I did the same with the Latest element as it did not move at all from its fixed position. I also fixed the width of the cards for small screens as they used to show in different sizes