Header issues fixed
I had an issue with the header for L screen. For some reason there was a blank space on the right side which extended the viewpoint and a scroller appeared on the bottom of the page for no reason as there was nothing there. I managed to sort it out. Works on IE too