Feedback fixes vol.2
Fixed the overflow of the newscard image on hover
Fixed some last issues with the position of the footers in the newscards as there were a few pixels difference on some screen sizes
Fixed the height of the main-photo-container div as it did not fill the whole screen
Fixed the position of the text in the photo-container for XL and L screen as it was not responsive. Fixed it for all other screens too including IE
Fixed the issue I have been having for ages with the submenu alignment to the left. It starts from the beginning of the navbar now on all screens it is visible
Left margins aligned with each other and navbar for the whole page