More fixes
Fixed the Latest font size and weight
Fixed all buttons as they were a bit too big
Fixed the font size in the culture part
Fixed the margin-bottom of newscards as there was none and the container was too close to the next element
Fixed the font-weight of the email-newsletter title. It was too bold
Fixed the Google cursor as it used to show as hand cursor but was not needed there. Also the logo was blurry in Chrome. I found a way to make it crisp and clear with transform property as the image itself was good quality.
Fixed the newscards on screen S when a single one is visible as they did not fill the screen. There was a width property that was not needed
Fixed the culture and subscribe buttons as their font-size was not correct