Fixes on my code
I fixed my SCSS code for awards, main-photo, newsletter, culture, cards and client-logos by making it looking good and added comments for better.
Also fixed the cursor issue with the Google logo for IE as it was showing as a hand/link cursor instead of default.
Fixed the issue with a blue border-shadow kind of thing appearing in IE on the top when clicked and the cursor used to change to the select (hand) one