Final bug fixes before submitting
Fixed an issue with the submenu. It was a bit too wide and a scroll button appeared on the bottom of the page. I changed 100vw with 100%
I added padding-left to the input element in newsletter
I added border-radius to the newscards
I added transition: none; to the button in cards as they used to have some kind of transition coming from somewhere else. I could not find where was it coming from and added this extra code
I fixed a few issues with the hero image for S and M screens. It did not fill the whole screen. I managed to fix that by changing the height calculation for those screens. It looks more adaptive to the screen size now. Also fixed the position of the text box to percentage in order to make it in the middle on all screens
Fixed the top-margin of the navbar-submenu in order to centre the text
I fixed the XL breakpoint to 1260px as it is on the original site. I had it on 1200px before
Fixed the position of btn-top for L screen so that it looks the same as the original site. There was misbehaving as the menu button was not aligned with the cyber security tab. I changed the position and made it aligned to it
Following Sam’s advice I decided to go through my code for the menu button and make sure it is done the way he suggested. I spent some time to find out how is it styled on the original website and managed to get it looks the same.