Fixes based on the feedback from Kayleigh
I added the style for the logo in the scss file
I removed the button tag from the HTML for the header support and contact buttons
I removed the button tag on all other occasions but left it in the navbar  as I could not find out how to make it work. Spent quite a lot of time but it got completely messed up if I remove the buttons
I made menu-lines a class as I have used it twice
I changed all span elements as per the instruction. I changed them with li elements
I removed the alt from the hero photo
I added alt to all places it was missing
I changed the figcaption with p and added margin 0 to make it look the same

I fixed a few h3 tags that were closed incorrectly when I checked my code in the validator
I did validate the code and there were no errors apart from the navbar issue
I fixed the distance between the buttons, search bar and menu button in the header for big screen sizes
I fixed the distance between the buttons in the headers for all other screen sizes
I managed to fix some alignment issues I was having on the L screen size. The logo and the rest of the buttons were responding differently to the screen size change. I got them to stay at the same place and move all together
I fixed the border radius on the search button on big screens
I fixed the slight different in the alignment between the navbar and the subnavbar on the left side
I fixed the box-shadow of the cards to reflect exactly the one of the main website
I fixed the space on top and under the cards as it was a bit too much on top. Made it the same size