Fixed the navbar HTML code and the borders of the navbar items
I managed to fix the styling of the navbar however the hover over effect is not there yet. Once hovered it shows the submenu but the colour on the main navbar is not showing
I spent some extra time as I could not find out why the main colour of the element did not change on hover. The code was there but was not working. I found out that after changing my HTML doc it must have changed the targeting in the style sheets. I changed it a bit by targeting a different element and managed to get it work
I fixed an issue that appeared on IE. It does not recognise space-evenly and shows the navbar completely messed up. I changed it to space-around for IE only in order to make it act in the same way as other browsers
I fixed the borders too. I added border to the left and right as first and last child and then added left and right to each element in the navbar. I made some changes to the border style too. Copied the style from the main website to reflect the borders in exactly the same way