Added sticky header and navbar
I added some JS to make the header and the navbar sticky. They disappear with scrolling down and appear again when user scrolls back up. I had issues with some of the elements on the page as the new positioning of the header and navbar misplaced my brand image and the header for some screens. I fixed the CSS in order to make it work and it all works fine on all browsers and screens