Slider fixed
I managed to get the dots on the screen and styled them as they are on the main website. I added some more styling overall and made them stay in the middle. I had some issues with IE as it did not show my images. I changed the path to links instead of actual files in order to make them work. Also had to fix some small issues with the cards container element as it got misplaced a bit. All looking good and working fine on all browsers now