Fixed all issues
Everything has been fixed.
I had a problem with the number of newscards shown on medium screen as they should be 2 but there was no way to hide the third because they all had the same class. I managed to add some extra php code and add a number after the class. Then I targeted the class in my scss file and it all works perfectly now