Fixed an issue with the newsletter checkbox
When I uploaded my work to cPanel there was an error and my form did not work. I had to change a few things and added the require files to the top of my contact.php.
After that the form did not submit correctly if the checkbox was not selected. I found out that I have to change my bindParam command and give it value. It was taking an integer by default which was not correct. I added that newsletter value should be a boolean and it works now.